Compute genomic prediction different methods

Compute genomic prediction different methods

Usage

compute_GP_methods(
  geno,
  pheno,
  traits,
  GP.method,
  nreps = 10,
  nfolds = 10,
  h = 1,
  nb.mtry = 10,
  nIter = 6000,
  burnIn = 1000,
  ntree = 100,
  p2d.temp = NULL,
  nb.cores = 1,
  p2f.stats = NULL
)

Arguments

geno

genomic data with genotypes in row (GID in rownames) and marker in columns. Values should be column centered and scaled.

pheno

phenotypic data with genotypes in row (in GID column) and traits in columns. Phenotypic value should be corrected for year and location effects beforehand.

traits

character vector of trait names

GP.method

character vector of genomic prediction methods to use

nreps

number of repetitions for cross-validation, default is 10

nfolds

number of folds for cross-validation, default is 10

h

bandwith parameter for RKHS, default is 1. If multiple values are provided, method will be RKHS Kernel Averaging.

nIter

number of iterations for RKHS, default is 6000

burnIn

number of burn-in iterations for RKHS, default is 1000

ntree

number of trees for RandomForest, default is 100

p2d.temp

path to directory to export temporary genomic prediction results, default is NULL (could cause error in parallelization if NULL).

nb.cores

number of cores to parallelize the computation, default is 1 (no parallelization)

p2f.stats

path to file to export genomic prediction results, default is NULL