Create table and file to add accession to T3

Create table and file to add accession to T3

Usage

create_accessions_T3(
  dat = NULL,
  checkDB = TRUE,
  p2f = NULL,
  return_table = TRUE,
  accession_name = NULL,
  population_name = NULL,
  `organization_name(s)` = NULL,
  `synonym(s)` = NULL,
  `variety(s)` = NULL,
  `country_of_origin(s)` = NULL,
  `notes(s)` = NULL,
  `accession_number(s)` = NULL,
  purdy_pedigree = NULL,
  filial_generation = NULL,
  species_name = "Triticum aestivum"
)

Arguments

dat

data frame with genotype information

checkDB

logical, whether to check the T3 database with BrAPI package to exclude accessions already in the database, default is TRUE.

p2f

path to file to write the output of the function

return_table

logical, whether to return the table or not, default is TRUE.

accession_name

character string, name of the column in dat that contains the accession name

population_name

character string, name of the column in dat that contains the population name

purdy_pedigree

character string, name of the column in dat that contains the purdy pedigree

filial_generation

character string, name of the column in dat that contains the filial generation

species_name

character string, name of the species, default is "Triticum aestivum" (fixed).

`organization_name(s)`

charac6er string, name of the column in dat that contains the organization name

`synonym(s)`

character string, name of the column in dat that contains the synonym name

`variety(s)`

character string, name of the column in dat that contains the variety name

`country_of_origin(s)`

character string, name of the column in dat that contains the country of origin

`notes(s)`

character string, name of the column in dat that contains the notes

`accession_number(s)`

character string, name of the column in dat that contains the accession number

Value

a data frame with the accessions to add to T3 if return_table is TRUE, otherwise a file is written to the path p2f

Author

Charlotte Brault