Create table and file to add phenotype data in T3

Create table and file to add phenotype data in T3

Usage

create_phenot_T3(
  dat = NULL,
  df_corresp_trait = NULL,
  p2f = NULL,
  return_table = TRUE,
  plot_name_col = NULL
)

Arguments

dat

data frame with phenotypic data, containing at least the plot name and the traits

df_corresp_trait

data frame with the correspondence between the column names in dat and the traits in T3, must countain the columns T3_trait and corresp_col.

p2f

path to file to write the output of the function

return_table

logical, whether to return the table or not, default is TRUE.

plot_name_col

character string, name of the column in dat that contains the plot name, default is "plot_name"

Value

a data frame with the phenotypes to add to T3 if return_table is TRUE, otherwise a file is written to the path p2f.

See also